Moved legacy converter file location since it doesn't actually have to do with serialization.
